added update function in NewsController.js

+++ server/src/controllers/NewsController.js
@@ -2,13 +2,12 @@ import db from '../models'
 import NotFoundError from '../errorHandlers/NotFoundError'
 import moment from 'moment'
 import {STATIC_PATH_MAIN_PHOTO_NEWS} from '../constants'
+import fs from "fs";
 const News = db.News;
 
 module.exports.createNews = async (req, res, next) => {
     try {
-        //console.log(JSON.parse(req.body.news));
-        //console.log(req.file);
-        const news = JSON.parse(req.body.news);
+        const news = req.body;
         news.main_img = req.file.filename;
         news.date = moment(news.date).format('YYYY-MM-DD');
         const result = await News.create(news);
@@ -40,22 +39,48 @@ module.exports.getAllNews = async (req, res, next) => {
 
 module.exports.getNews = async (req, res, next) => {
     const news = await News.findByPk(req.params.id);
-    news.main_img = STATIC_PATH_MAIN_PHOTO_NEWS + news.main_img
+    news.main_img = STATIC_PATH_MAIN_PHOTO_NEWS + news.main_img;
     if(!news){
         return next(new NotFoundError("News not found"));
     }
     res.send(news);
 };
 
+module.exports.update = async (req, res, next) => {
+    try {
+        let news = req.body;
+        const oldNews = await News.findByPk(req.params.id);
+        if (!oldNews) {
+            return next(new NotFoundError('Post was not found'));
+        }
+        if (req.file) {
+            fs.unlink("public/images/mainPhotoNews/" + oldNews.main_img, (err) => {
+                if (err) throw err
+            });
+            news.main_img = req.file.filename;
+        }
+        if(news.date) {
+            news.date = moment(news.date).format('YYYY-MM-DD');
+        }
+        await News.update(news, {where: {id: req.params.id}});
+        res.status(201).send();
+    } catch (e) {
+        next(e);
+    }
+};
+
 module.exports.deleteNews = async (req, res, next) => {
     try {
-        const countOfModifiedFields = await News.destroy({where: {id: req.params.id}});
-        if(countOfModifiedFields < 1){
-            return next(new NotFoundError("News not found"));
+        const news = await News.findByPk(req.params.id);
+        if(!news){
+                return next(new NotFoundError('News not found'))
         }
+        fs.unlink("public/images/mainPhotoNews/" + news.main_img, (err) => {
+            if (err) throw err
+        });
+        const countOfModifiedFields = await News.destroy({where: {id: req.params.id}});
         res.send({modified: countOfModifiedFields});
     } catch(e){
         return next(e);
     }
-
 };
